dspy/rag-agent/requirements.txt
pydantic>=2.5.0,<2.12
dspy==2.6.27
dspy-ai==2.6.27
mlflow[databricks]==3.1
databricks-agents>=1.2.0
databricks-connect>=16.1.0
databricks-vectorsearch
databricks-sdk
openai<2
python-dotenv
uv